# perf: remove unused ASTNode parsing