--- a/setup.py
+++ b/setup.py
@@ -9,6 +9,7 @@ extras_require = {
         "pytest-cov",  # Coverage analyzer plugin
         "pytest-mock",  # For creating and using mocks
         "hypothesis>=6.2.0,<7.0",  # Strategy-based fuzzer
+        "pytest-benchmark",  # For performance tests
     ],
     "lint": [
         "black>=23.12.0,<24",  # Auto-formatter and linter
@@ -21,6 +22,7 @@ extras_require = {
         "mdformat>=0.7.17",  # Auto-formatter for markdown
         "mdformat-gfm>=0.3.5",  # Needed for formatting GitHub-flavored markdown
         "mdformat-frontmatter>=0.4.1",  # Needed for frontmatters-style headers in issue templates
+        "mdformat-pyproject>=0.0.1",  # Allows configuring in pyproject.toml
     ],
     "doc": [
         "Sphinx>=6.1.3,<7",  # Documentation generator