### What I did

<!-- The `fixes:` field denotes an issue that will be marked resolved by merging this PR -->

fixes: #https://github.com/ApeWorX/ape/issues/1726

### How I did it

* Del ASTNode logic for now as there are no tracing features yet in solidity so this does not affect anything and it was too pre-emptive to add anyway. I would like to get tracing features working and somehing like this is a first step, but the performance concerns leads me to think are going to probably want the raw dicts in the manifest instead and parse them later so compiling alone does not take forever, but will probably adjust `ape-vyper` like that first.

### How to verify it

With this PR:

```
(ape310) ➜  ape-playground git:(repro1726) ✗ time ape compile -f
INFO: Compiling 'token.sol'.
INFO: Compiling 'Lib.sol'.
INFO (ape-solidity): Compiling using Solidity compiler '0.8.23+commit.f704f362'.
ape compile -f  4.63s user 0.76s system 122% cpu 4.391 total
```

Without this PR:

```
(ape310) ➜  ape-playground git:(repro1726) ✗ time ape compile -f
INFO: Compiling 'token.sol'.
INFO: Compiling 'Lib.sol'.
INFO (ape-solidity): Compiling using Solidity compiler '0.8.23+commit.f704f362'.
ape compile -f  33.43s user 0.84s system 102% cpu 33.307 total
```

4.63 seconds with PR, 33.43 w

**That is 86.8% increase!**

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)