@antazoey 2024-01-05
I couldn't get the regression testing portion figured out... It is stupidly complicated I guess and I can't really find a decent example.

For now will have to rely on looking at the output from past CI runs...